requirements.txt
pyyaml>=6.0
requests>=2.28
numpy>=1.24
pandas>=2.0
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12
tqdm>=4.65
